requirements.txt
Pillow>=10.4.0
streamlit==1.34.0
ultralytics==8.2.8
pymupdf==1.21.1
torch
opencv-python
numpy